requirement.txt
pandas
openai>=1.2.4,<1.7.0
streamlit
python-dotenv
requests
langchain
langchain-community
langchain-text-splitters
sentence-transformers
chromadb